Migrate all entry scirpts to TS.
