+++ package.json
@@ -11,11 +11,13 @@
     "yaml": "^2.2.1"
   },
   "devDependencies": {
+    "@jest/globals": "^29.7.0",
     "@tsconfig/recommended": "^1.0.3",
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^20.11.5",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3",
-    "jest": "^29.7.0"
+    "typescript": "^5.3.3"
   }
 }
